% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_methods.R
\docType{methods}
\name{tiff_header}
\alias{tiff_header}
\alias{tiff_header,TiffArraySeed-method}
\alias{tiff_header,TiffArray-method}
\alias{tiff_header,TiffArrayList-method}
\alias{tiff_header,DelayedArray-method}
\alias{tiff_header,HDF5Array-method}
\alias{tiff_header,HDF5ArraySeed-method}
\alias{tiff_header,ANY-method}
\title{Dump a Tiff header}
\usage{
tiff_header(image)

\S4method{tiff_header}{TiffArray}(image)

\S4method{tiff_header}{TiffArrayList}(image)

\S4method{tiff_header}{DelayedArray}(image)

\S4method{tiff_header}{HDF5Array}(image)

\S4method{tiff_header}{HDF5ArraySeed}(image)

\S4method{tiff_header}{ANY}(image)

\S4method{tiff_header}{TiffArraySeed}(image)
}
\arguments{
\item{image}{An image or TiffArray object.}
}
\value{
A list of class \code{tiffHeader}, which has
the attribute information.
}
\description{
Dump a Tiff header
}
\examples{
nii_fname = system.file("img", "Rlogo.tif", package = "ijtiff")
tiff_header(nii_fname)
res = writeTiffArray(nii_fname)
tiff_header(res)
}
