% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiffArray.R, R/aaa_TiffArraySeed.R,
%   R/zzz_TiffMatrix.R
\docType{class}
\name{TiffArray}
\alias{TiffArray}
\alias{class:TiffArray}
\alias{TiffArray-class}
\alias{coerce}
\alias{coerce,ANY,TiffArray-method}
\alias{coerce,DelayedArray,TiffArray-method}
\alias{coerce,DelayedMatrix,TiffMatrix-method}
\alias{coerce,HDF5Array,TiffArray-method}
\alias{coerce,HDF5Array,TiffMatrix-method}
\alias{coerce,HDF5Matrix,TiffMatrix-method}
\alias{coerce,TiffArray,TiffMatrix-method}
\alias{coerce,TiffArrayList,TiffMatrix-method}
\alias{coerce,numeric,TiffMatrix-method}
\alias{coerce,numeric,TiffArray-method}
\alias{coerce,TiffArray,ijtiff_img-method}
\alias{coerce,TiffMatrix,ijtiff_img-method}
\alias{coerce,TiffMatrix,TiffArray-method}
\alias{coerce,ANY,TiffMatrix-method}
\alias{coerce,TiffArrayList,TiffArray-method}
\title{Tiff images as DelayedArray objects, using HDF5Array}
\usage{
TiffArray(filepath, name = "image", type = NA, header = NULL)
}
\arguments{
\item{filepath}{The path (as a single character string) to the HDF5
file where the dataset is located.}

\item{name}{The name of the image in the HDF5 file.}

\item{type}{\code{NA} or the R atomic type, passed to
\code{\link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array()}}.}

\item{header}{list of attribute information to override call of
\link{tiff_header}}
}
\value{
A \code{TiffArray} object
}
\description{
TiffArray, a \link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array} with a header.
}
\note{
All the operations available for
\link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array} objects work on
TiffArray objects.
}
\examples{
nii_fname = system.file("img", "Rlogo.tif", package = "ijtiff")
res = TiffArray(nii_fname)
res2 = TiffArray(slot(slot(res, "seed"), "filepath"))
res2 = TiffArray(slot(res, "seed"))
}
