% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiffArrayList.R
\name{TiffArrayList}
\alias{TiffArrayList}
\title{Write TiffArray Object}
\usage{
TiffArrayList(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{a character or list of objects to be called with
\link{writeTiffArray}}

\item{...}{additional arguments to pass to \link{TiffArray}}

\item{verbose}{show progress bars if \code{pbapply} package installed.}
}
\value{
A list of class \code{TiffArrayList}, which is a
list if \code{TiffArray} objects
}
\description{
Write TiffArray Object
}
\examples{
nii_fname = system.file("img", "2ch_ij.tif", package = "ijtiff")
nii_fname = rep(nii_fname, 3)
res = TiffArrayList(nii_fname)
if (requireNamespace("pbapply", quietly = TRUE)) {
   res = TiffArrayList(nii_fname, verbose = TRUE)
} else {
   testthat::expect_warning({
   res = TiffArrayList(nii_fname, verbose = TRUE)
   })
}
testthat::expect_is(res, "TiffArrayList")
mat = as(res, "TiffMatrix")
arr = as(res, "TiffArray")
h5 = unlist(DelayedArray::seedApply(res, slot, "filepath"))
res2 = TiffArrayList(h5)
testthat::expect_is(res2, "TiffArrayList")
mat = as(res2, "TiffMatrix")
arr = as(res2, "TiffArray")
}
