% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_TiffArraySeed.R
\docType{class}
\name{TiffArraySeed-class}
\alias{TiffArraySeed-class}
\alias{DelayedArray,TiffArraySeed-method}
\alias{TiffArraySeed}
\title{Seed for TiffArray Class}
\usage{
TiffArraySeed(filepath, name = "image", type = NA, header = NULL)
}
\arguments{
\item{filepath}{The path (as a single character string) to the HDF5
file where the dataset is located.}

\item{name}{The name of the image in the HDF5 file.}

\item{type}{\code{NA} or the R atomic type, passed to
\code{\link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array()}}}

\item{header}{list of header information to override call of
\link{tiff_header}}
}
\value{
A \code{TiffArraySeed} object
}
\description{
Seed for TiffArray Class
}
\examples{
nii_fname = system.file("img", "2ch_ij.tif", package = "ijtiff")
res = TiffArraySeed(nii_fname)
hdr = tiff_header(res)
res2 = TiffArraySeed(nii_fname, header = hdr)
}
