% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_TiffArraySeed.R, R/zzz_TiffMatrix.R
\docType{class}
\name{TiffMatrix-class}
\alias{TiffMatrix-class}
\alias{matrixClass}
\alias{matrixClass,TiffArray-method}
\title{TiffMatrix Class}
\usage{
\S4method{matrixClass}{TiffArray}(x)
}
\arguments{
\item{x}{Typically a DelayedArray object.}
}
\value{
A \code{TiffMatrix} object.
}
\description{
TiffMatrix Class
}
