% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTiffArray.R
\name{writeTiffArray}
\alias{writeTiffArray}
\title{Write TiffArray Object}
\usage{
writeTiffArray(x, filepath = tempfile(fileext = ".h5"), name = "image",
  chunkdim = NULL, level = NULL, verbose = FALSE, header = NULL,
  overwrite = FALSE)
}
\arguments{
\item{x}{a \code{tiff} object or file path to \code{tiff} file}

\item{filepath}{The path (single character string) to the HDF5 file.}

\item{name}{The name of the image in the HDF5 file.}

\item{chunkdim}{The dimensions of the chunks to use for
writing the data to disk.
Passed to \link[HDF5Array:writeHDF5Array]{HDF5Array::writeHDF5Array}.}

\item{level}{The compression level,
passed to \link[HDF5Array:writeHDF5Array]{HDF5Array::writeHDF5Array}.}

\item{verbose}{Display progress,
passed to \link[HDF5Array:writeHDF5Array]{HDF5Array::writeHDF5Array}.}

\item{header}{list of header information;
overrides call of \link{tiff_header}}

\item{overwrite}{\code{FALSE} by default and an in the
event that an HDF5 file already exists for \code{filepath}
input then do not overwrite it.
If set to \code{TRUE} then the "image" and "hdr" objects at this file
location will overwrite.}

\item{header_name}{The name of the header in the HDF5 file.}
}
\value{
A \code{TiffArray} object.
}
\description{
Write TiffArray Object
}
\examples{
nii_fname = system.file("img", "Rlogo.tif", package = "ijtiff")
res = writeTiffArray(nii_fname)
tiff_header(res)
filepath = tempfile(fileext = ".h5")
res = writeTiffArray(nii_fname, filepath = filepath)
testthat::expect_error(
   writeTiffArray(nii_fname, filepath = filepath),
   regexp = "already exist",
)
res = writeTiffArray(nii_fname, filepath = filepath, overwrite = TRUE)
img = ijtiff::read_tif(nii_fname)
writeTiffArray(c(img), header = tiff_header(img))
}
